utils: move tristate into utils
