+++ src/v/serde/rw/tristate_rw.h
@@ -10,7 +10,7 @@
 #pragma once
 
 #include "serde/rw/rw.h"
-#include "tristate.h"
+#include "utils/tristate.h"
 
 namespace serde {
