+++ src/v/storage/types.h
@@ -21,7 +21,7 @@
 #include "storage/fwd.h"
 #include "storage/key_offset_map.h"
 #include "storage/scoped_file_tracker.h"
-#include "tristate.h"
+#include "utils/tristate.h"
 
 #include <seastar/core/abort_source.hh>
 #include <seastar/core/file.hh> //io_priority
