+++ src/v/utils/tests/tristate_test.cc
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "tristate.h"
+#include "utils/tristate.h"
 
 #include <boost/test/unit_test.hpp>
