+++ src/v/utils/tristate.h
